requirements.txt
flask==3.0.3
requests==2.32.3
schedule==1.2.2
gunicorn==22.0.0
psycopg2-binary==2.9.9
